moving cluster active logic to resource_abstractor
